# shared/openapi.json
{
  "components": {
    "schemas": {
...
      "AnalyzeRequest": {
        "properties": {
          "platform": {
            "title": "Platform",
            "type": "string"
          },
          "priceHistory": {
            "items": {
              "$ref": "#/components/schemas/PriceHistoryPoint"
            },
            "title": "Pricehistory",
            "type": "array"
          },
          "product": {
            "$ref": "#/components/schemas/Product"
          },
          "reviews": {
            "items": {
              "$ref": "#/components/schemas/Review"
            },
            "title": "Reviews",
            "type": "array"
          },
          "session": {
            "$ref": "#/components/schemas/SessionContext"
          },
          "userBudget": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UserBudget"
              },
              {
                "type": "null"
              }
            ]
          },
          "userId": {
            "title": "Userid",
            "type": "string"
          }
        },
        "required": [
          "userId",
          "platform",
          "product",
          "session"
        ],
        "title": "AnalyzeRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Türk e-ticaret için 5 ajanlı satın alma analiz servisi. Fiyat geçmişi crowdsource veritabanından beslenir.",
    "title": "Thundrly — Backend",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/analyze-purchase": {
      "post": {
        "description": "Ürün, yorum, fiyat geçmişi, bütçe ve oturum bilgisini alır; yeşil/sarı/kırmızı bir karar ve 3 gerekçe döner. `priceHistory` boş ise sunucu, ürün URL'sine ait crowdsource veritabanından son 90 günü çeker.",
        "operationId": "analyze_purchase_api_analyze_purchase_post",
        "parameters": [
          {
            "description": "True olduğunda Gemini yanıt cache'i atlanır — kullanıcı panelden 'Yeniden analiz et' istediğinde gönder.",
            "in": "query",
            "name": "force_refresh",
            "required": false,
            "schema": {
              "default": false,
              "description": "True olduğunda Gemini yanıt cache'i atlanır — kullanıcı panelden 'Yeniden analiz et' istediğinde gönder.",
              "title": "Force Refresh",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest",
                "examples": {
                  "green": {
                    "summary": "Gerçek indirim + planlı + bütçe içinde",
                    "value": {
                      "platform": "n11-demo",
                      "priceHistory": [
                        {
                          "date": "2026-02-01",
                          "price": 160
                        },
                        {
                          "date": "2026-03-01",
                          "price": 155
                        },
                        {
                          "date": "2026-04-01",
                          "price": 165
                        },
                        {
                          "date": "2026-05-01",
                          "price": 180
                        },
                        {
                          "date": "2026-05-14",
                          "price": 145
                        }
                      ],
                      "product": {
                        "category": "Kitap",
                        "currency": "TRY",
                        "imageUrl": "https://demo.local/img/sapiens.jpg",
                        "originalPrice": 180,
                        "price": 145,
                        "rating": 4.8,
                        "reviewCount": 5230,
                        "title": "Sapiens: Hayvanlardan Tanrılara",
                        "url": "https://demo.local/product/sapiens"
                      },
                      "reviews": [
                        {
                          "date": "2026-03-12",
                          "rating": 5,
                          "text": "Yuval Noah Harari'nin bakış açısı çok geniş, akıcı bir kitap."
                        },
                        {
                          "date": "2026-03-18",
                          "rating": 5,
                          "text": "İnsanlık tarihini tek kitapta toparlayan harika bir eser."
                        },
                        {
                          "date": "2026-04-02",
                          "rating": 4,
                          "text": "Bazı bölümler tartışmaya açık ama düşündürücü."
                        },
                        {
                          "date": "2026-04-22",
                          "rating": 5,
                          "text": "Çevirisi başarılı, kağıt kalitesi iyi."
                        }
                      ],
                      "session": {
                        "clickSpeedMs": 2400,
                        "currentHour": 16,
                        "purchasesToday": 0,
                        "searchedBefore": true,
                        "timeOnPageSeconds": 210
                      },
                      "userBudget": {
                        "categoryLimit": 500,
                        "categorySpent": 80,
                        "currency": "TRY",
                        "monthlyLimit": 2500,
                        "monthlySpent": 600
                      },
                      "userId": "demo-user"
                    }
                  },
                  "red": {
                    "summary": "Şüpheli indirim + bütçe aşımı + dürtüsel",
                    "value": {
                      "platform": "trendyol-demo",
                      "priceHistory": [
                        {
                          "date": "2026-04-15",
                          "price": 780
                        },
                        {
                          "date": "2026-04-22",
                          "price": 780
                        },
                        {
                          "date": "2026-04-29",
                          "price": 820
                        },
                        {
                          "date": "2026-05-05",
                          "price": 1650
                        },
                        {
                          "date": "2026-05-10",
                          "price": 1650
                        },
                        {
                          "date": "2026-05-14",
                          "price": 990
                        }
                      ],
                      "product": {
                        "category": "Giyim",
                        "currency": "TRY",
                        "imageUrl": "https://demo.local/img/hoodie.jpg",
                        "originalPrice": 1650,
                        "price": 990,
                        "rating": 4.7,
                        "reviewCount": 842,
                        "title": "Oversize Siyah Hoodie",
                        "url": "https://demo.local/product/hoodie"
                      },
                      "reviews": [
                        {
                          "date": "2026-05-10",
                          "rating": 5,
                          "text": "Çok güzel ürün hızlı kargo mükemmel kalite"
                        },
                        {
                          "date": "2026-05-10",
                          "rating": 5,
                          "text": "Çok güzel ürün hızlı kargo kalite süper"
                        },
                        {
                          "date": "2026-05-11",
                          "rating": 5,
                          "text": "Mükemmel kalite hızlı kargo tavsiye ederim"
                        },
                        {
                          "date": "2026-05-11",
                          "rating": 5,
                          "text": "Çok güzel kalite hızlı kargo"
                        },
                        {
                          "date": "2026-05-11",
                          "rating": 5,
                          "text": "Mükemmel ürün hızlı kargo süper kalite"
                        },
                        {
                          "date": "2026-05-12",
                          "rating": 5,
                          "text": "Tam istediğim gibi geldi tavsiye ederim"
                        },
                        {
                          "date": "2026-05-12",
                          "rating": 5,
                          "text": "Çok güzel ürün hızlı kargo mükemmel"
                        },
                        {
                          "date": "2026-04-15",
                          "rating": 2,
                          "text": "Beden uymadı, kumaş beklediğim gibi değildi."
                        }
                      ],
                      "session": {
                        "clickSpeedMs": 420,
                        "currentHour": 23,
                        "purchasesToday": 2,
                        "searchedBefore": false,
                        "timeOnPageSeconds": 18
                      },
                      "userBudget": {
                        "categoryLimit": 1000,
                        "categorySpent": 1700,
                        "currency": "TRY",
                        "monthlyLimit": 3000,
                        "monthlySpent": 2400
                      },
                      "userId": "demo-user"
                    }
                  },
                  "yellow": {
                    "summary": "Kısmen manipüle fiyat + sınırda bütçe",
                    "value": {
                      "platform": "hepsiburada-demo",
                      "priceHistory": [
                        {
                          "date": "2026-04-15",
                          "price": 1199
                        },
                        {
                          "date": "2026-04-25",
                          "price": 1199
                        },
                        {
                          "date": "2026-05-01",
                          "price": 1899
                        },
                        {
                          "date": "2026-05-08",
                          "price": 1899
                        },
                        {
                          "date": "2026-05-14",
                          "price": 1450
                        }
                      ],
                      "product": {
                        "category": "Elektronik",
                        "currency": "TRY",
                        "imageUrl": "https://demo.local/img/headphones.jpg",
                        "originalPrice": 1899,
                        "price": 1450,
                        "rating": 4.4,
                        "reviewCount": 312,
                        "title": "Kablosuz Kulaklık",
                        "url": "https://demo.local/product/headphones"
                      },
                      "reviews": [
                        {
                          "date": "2026-05-09",
                          "rating": 5,
                          "text": "Ses kalitesi beklediğimden iyi, gürültü engelleme başarılı."
                        },
                        {
                          "date": "2026-05-10",
                          "rating": 4,
                          "text": "Pil ömrü iyi ama kulaklık biraz sıkıyor."
                        },
                        {
                          "date": "2026-05-11",
                          "rating": 5,
                          "text": "Çok güzel ürün hızlı kargo."
                        },
                        {
                          "date": "2026-05-12",
                          "rating": 5,
                          "text": "Hızlı kargo çok güzel mükemmel kalite."
                        },
                        {
                          "date": "2026-05-13",
                          "rating": 3,
                          "text": "Mikrofon kalitesi orta, ses kalitesi iyi."
                        }
                      ],
                      "session": {
                        "clickSpeedMs": 900,
                        "currentHour": 22,
                        "purchasesToday": 1,
                        "searchedBefore": false,
                        "timeOnPageSeconds": 28
                      },
                      "userBudget": {
                        "categoryLimit": 1500,
                        "categorySpent": 700,
                        "currency": "TRY",
                        "monthlyLimit": 3500,
                        "monthlySpent": 2100
                      },
                      "userId": "demo-user"
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyzeResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Ürün satın alımını 5 ajanla analiz et"
      }
    },
    "/api/analyze-purchase/stream": {
      "post": {
        "description": "Her satır bağımsız bir JSON nesnesidir (`application/x-ndjson`). Sıra: 4 paralel sinyal ajanı için `node_finished`, ardından decision için `node_finished`, en sonda `verdict` (tam AnalyzeResponse). Hata olursa son satır `error` olur. Karar mantığı normal endpoint ile aynıdır; tek fark stream şeklidir.",
        "operationId": "analyze_purchase_stream_api_analyze_purchase_stream_post",
        "parameters": [
          {
            "in": "query",
            "name": "force_refresh",
            "required": false,
            "schema": {
              "default": false,
              "title": "Force Refresh",
              "type": "boolean"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyzeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Aynı analiz, ajan tamamlanma olaylarını NDJSON ile streamler"
      }
    },
    "/api/cache/purge": {
      "post": {
        "description": "Hedefli temizleme: ?userId=... verilirse o kullanıcıya ait, ?url=... verilirse o URL'ye ait kayıtlar düşürülür. Hiçbiri verilmezse tüm cache temizlenir. THUNDRLY_ADMIN_TOKEN gerekli (tanımlı değilse açık).",
        "operationId": "cache_purge_api_cache_purge_post",
        "parameters": [
          {
            "in": "query",
            "name": "userId",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Userid"
            }
          },
          {
            "in": "query",
            "name": "url",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Url"
            }
          },
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cache Purge Api Cache Purge Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Gemini cache içeriğini boşalt (admin)"
      }
    },
    "/api/cache/stats": {
      "get": {
        "description": "Bellek içi Gemini response cache'inin sayaçlarını döner. Hit oranı, evictions ve invalidations metriklerini gözlemlemek için kullan. THUNDRLY_ADMIN_TOKEN tanımlıysa Authorization header'ında geçirilmesi gerekir; tanımlı değilse herkese açık.",
        "operationId": "cache_stats_api_cache_stats_get",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "title": "Response Cache Stats Api Cache Stats Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Gemini cache hit/miss istatistikleri"
      }
    },
...
  }
}